refactor: extract Go/Maven/Cargo/uvx parsers into parsers/compiled_parsers.py
Move _parse_go_mod_requires, parse_go_packages, parse_maven_packages,
parse_cargo_packages, and detect_uvx_package into dedicated module.
Re-export from __init__.py for backward compatibility.

parsers/__init__.py: 694 → 430 lines (-38%)
Total reduction from original: 1,250 → 430 lines (-66%)
